+++ pyriemann/channelselection.py
@@ -14,12 +14,12 @@ class ElectrodeSelection(BaseEstimator, TransformerMixin):
     on the maximization of the distance between centroids. This is done by a
     backward elimination where the electrode that carries the less distance is
     removed from the subset at each iteration.
-    This algorith is described in [1]_.
+    This algorithm is described in [1]_.
 
     Parameters
     ----------
     nelec : int (default 16)
-        the number of electrode to keep in the final subset.
+        The number of electrode to keep in the final subset.
     metric : string | dict (default: 'riemann')
         The type of metric used for centroid and distance estimation.
         see `mean_covariance` for the list of supported metric.
@@ -66,19 +66,22 @@ class ElectrodeSelection(BaseEstimator, TransformerMixin):
 
         Parameters
         ----------
-        X : ndarray, shape (n_trials, n_channels, n_channels)
-            ndarray of SPD matrices.
-        y : ndarray shape (n_trials, 1)
-            labels corresponding to each trial.
-        sample_weight : None | ndarray shape (n_trials, 1)
-            the weights of each sample. if None, each sample is treated with
-            equal weights.
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of SPD matrices.
+        y : None | ndarray, shape (n_matrices, 1)
+            Labels corresponding to each matrix.
+        sample_weight : None | ndarray, shape (n_matrices, 1)
+            Weights corresponding to each matrix.
+            If None, each matrix is treated with equal weights.
 
         Returns
         -------
         self : ElectrodeSelection instance
             The ElectrodeSelection instance.
         """
+        if y is None:
+            y = np.ones((X.shape[0]))
+
         mdm = MDM(metric=self.metric, n_jobs=self.n_jobs)
         mdm.fit(X, y, sample_weight=sample_weight)
         self.covmeans_ = mdm.covmeans_
@@ -109,13 +112,13 @@ class ElectrodeSelection(BaseEstimator, TransformerMixin):
 
         Parameters
         ----------
-        X : ndarray, shape (n_trials, n_channels, n_channels)
-            ndarray of SPD matrices.
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of SPD matrices.
 
         Returns
         -------
-        covs : ndarray, shape (n_trials, n_elec, n_elec)
-            The covariances matrices after reduction of the number of channels.
+        covs : ndarray, shape (n_matrices, n_elec, n_elec)
+            Set of SPD matrices after reduction of the number of channels.
         """
         return X[:, self.subelec_, :][:, :, self.subelec_]
 
@@ -125,7 +128,7 @@ class FlatChannelRemover(BaseEstimator, TransformerMixin):
 
     Attributes
     ----------
-    channels : ndarray, shape (n_good_channels)
+    channels_ : ndarray, shape (n_good_channels)
         The indices of the non-flat channels.
     """
 
@@ -134,15 +137,15 @@ class FlatChannelRemover(BaseEstimator, TransformerMixin):
 
         Parameters
         ----------
-        X : ndarray, shape (n_trials, n_channels, n_times)
-            Training data.
-        y : ndarray, shape (n_trials, n_dims) | None, optional
+        X : ndarray, shape (n_matrices, n_channels, n_times)
+            Multi-channel time-series.
+        y : ndarray, shape (n_matrices, n_dims) | None, optional
             The regressor(s). Defaults to None.
 
         Returns
         -------
-        X : ndarray, shape (n_trials, n_good_channels, n_times)
-            The data without flat channels.
+        X : ndarray, shape (n_matrices, n_good_channels, n_times)
+            Multi-channel time-series without flat channels.
         """
         std = np.mean(np.std(X, axis=2) ** 2, 0)
         self.channels_ = np.where(std)[0]
@@ -153,13 +156,13 @@ class FlatChannelRemover(BaseEstimator, TransformerMixin):
 
         Parameters
         ----------
-        X : ndarray, shape (n_trials, n_channels, n_times)
-            Training data.
+        X : ndarray, shape (n_matrices, n_channels, n_times)
+            Multi-channel time-series.
 
         Returns
         -------
-        X : ndarray, shape (n_trials, n_good_channels, n_times)
-            The data without flat channels.
+        X : ndarray, shape (n_matrices, n_good_channels, n_times)
+            Multi-channel time-series without flat channels.
         """
         return X[:, self.channels_, :]
 
@@ -168,15 +171,15 @@ class FlatChannelRemover(BaseEstimator, TransformerMixin):
 
         Parameters
         ----------
-        X : ndarray, shape (n_trials, n_channels, n_times)
-            Training data.
-        y : ndarray, shape (n_trials, n_dims) | None, optional
+        X : ndarray, shape (n_matrices, n_channels, n_times)
+            Multi-channel time-series.
+        y : ndarray, shape (n_matrices, n_dims) | None, optional
             The regressor(s). Defaults to None.
 
         Returns
         -------
-        X : ndarray, shape (n_trials, n_good_channels, n_times)
-            The data without flat channels.
+        X : ndarray, shape (n_matrices, n_good_channels, n_times)
+            Multi-channel time-series without flat channels.
         """
         self.fit(X, y)
         return self.transform(X)