+++ doc/whatsnew.rst
@@ -44,7 +44,7 @@ v0.2.8.dev
 
 - Add Riemannian Kernel Function: :func:`pyriemann.utils.kernel.kernel_riemann`
 
-- Add Riemannian Support Vector Machine: :class:`pyriemann.classification.SVC` and :class:`pyriemann.regression.SVR`
+- Fix ``fit`` in :class:`pyriemann.channelselection.ElectrodeSelection`
 
 
 v0.2.7 (June 2021)